pull up DuplicateLimitAfterMvExpand
